+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Fixed warning "Added non-passive event listener..." in Chrome caused by CodeMirror (https://github.com/codemirror/codemirror5/issues/6735)
 - Removed missed `console.log()` in progressbar
 
 ## 1.0.0-beta.86 (27-09-2024)